K2: Use KotlinScopeEnlarger for scope provider
K2 completion does use KotlinScopeEnlarger for the GlobalSearchScope.
As a result, the KotlinScopeEnlarger registered by Android Studio is not
triggered when app developers use the completion for app classes
generated from resources whose scope is supposed to be added by the
KotlinScopeEnlarger. This commit enlarges the search scope inside
`IdeKotlinByModulesResolutionScopeProvider` with the registered
`KotlinScopeEnlarger`.

^KTIJ-30502